+++ mnist_loader.h
@@ -5,6 +5,6 @@
 class mnist_loader {
 public:
     static int reverse_int(int i);
-    static std::vector<std::vector<double>> load_images(const std::string &filename);
+    static std::vector<std::vector<float>> load_images(const std::string &filename);
     static std::vector<int> load_labels(const std::string &filename);
 };
